057 - Introducao a testes, utilizando o RestAssured e JUnit

+++ algafood-api/src/main/java/com/algaworks/algafood/domain/service/CozinhaService.java
@@ -36,6 +36,9 @@ public class CozinhaService {
 		return cozinhaRepository.findById(id).orElseThrow(()-> new CozinhaNaoEncontradaException(id));
 	}
 	public void remover(Long id) {
+		if (!cozinhaRepository.existsById(id))
+			throw new CozinhaNaoEncontradaException(id);
+
 		try {
 			cozinhaRepository.deleteById(id);	
 		} catch (DataIntegrityViolationException e) {
